API Structure and Design

API Structure and Design

The API is segregated into 3 modules:

Common

Cos

PD

Common module has general system access and file access and parsing APIs. The ParserState type has been taken from the JuliaIO/JSON.jl. The file headers are not added. Hence, author acknowledges the efforts of the developers for the same package and expects the same be honored by any person developing any derivative work. Note-ParserState is no longer in use. The parser has been moved to the BufferedStreams interfaces. Some minor helper methods are ported to the new interface.

Cos module is the low level file format for PDF. Carousel Object Structure was original term proposed inside Adobe which later transformed into Acrobat. Cos layer has the object structure, definition and the cross references to access them.

PD module is the higher level document access layer. Accessing PDF pages or extracting the content from there or understanding document rendering using fonts or image objects will be typically in this layer. Please note that many objects in the PD layer actually refer to the Cos structure. You can consider PD Layer as the business logic while Cos Layer as the database for it.

Common

PDFIO.Common.CDTextString — Type.

    CDTextString

PDF file format structure provides two primary string types. Hexadecimal string CosXString and literal string CosLiteralString. However, these are mere binary representation of string types without having any encoding associated for semantic representation. Determination of encoding is carried out mostly by associated fonts and character maps in the content stream. There are also strings used in descriptions and other attributes of a PDF file where no font or mapping information is provided. This represents the string type in such situations. Typically, strings in PDFs are 3 types.

Text string a. PDDocEncoded string - Similar to ISO_8859-1 b. UTF-16BE strings

ASCII string

Byte string - Pure binary data no interpretation

1 and 2 can be represented by the CDTextString. convert methods are provided to translate the CosString to CDTextString

source

PDFIO.Common.CDDate — Type.

    CDDate

Internally represented as string objects, these are timezone enabled date and time objects.

PDF files support the string format: (D:YYYYMMDDHHmmSSOHH'mm)

source

PDFIO.Common.CDDate — Method.

    CDDate(s::CDTextString)

PDF files support the string format: (D:YYYYMMDDHHmmSSOHH'mm)

source

PDFIO.Common.CDRect — Type.

    CDRect

An CosArray representation of a rectangle in the lower left and upper right point format

source

COS Objects

PDFIO.Cos.CosObject — Type.

    CosObject

PDF is a structured document format with lots of internal data structures like dictionaries, arrays, trees. CosObject is the interface to access these objects and get detailed access to the objects and gather additional information. Although, defined in the COS layer, objects of these type are returned from almost all the APIs. Hence, the objects have a separate significance whether you need to use the Cos layer or not. Below is the object hierarchy.

CosObject                           Abstract
    CosNull                         Value (CosNullType)
CosString                           Abstract
CosName                             Concrete
CosNumeric                          Abstract
    CosInt                          Concrete
    CosFloat                        Concrete
CosBoolean                          Concrete
    CosTrue                         Value (CosBoolean)
    CosFalse                        Value (CosBoolean)
CosDict                             Concrete
CosArray                            Concrete
CosStream                           Concrete (always wrapped as an indirect object)
CosIndirectObjectRef                Concrete (only useful when CosDoc is available)

source

PDFIO.Cos.CosNull — Constant.

    CosNull

source

PDFIO.Cos.CosString — Type.

    CosString

source

PDFIO.Cos.CosName — Type.

    CosName

source

PDFIO.Cos.@cn_str — Macro.

    @cn_str

source

PDFIO.Cos.CosNumeric — Type.

    CosNumeric

source

PDFIO.Cos.CosInt — Type.

    CosFloat

source

PDFIO.Cos.CosFloat — Type.

    CosFloat

source

PDFIO.Cos.CosBoolean — Type.

    CosBoolean

source

PDFIO.Cos.CosDict — Type.

    CosDict

source

PDFIO.Cos.CosArray — Type.

    CosArray

source

PDFIO.Cos.CosStream — Type.

    CosStream

source

PDFIO.Cos.CosIndirectObjectRef — Type.

    CosIndirectObjectRef

A parsed data structure to ensure the object information is stored as an object. This has no meaning without a associated CosDoc. When a reference object is hit the object should be searched from the CosDoc and returned.

source

PD

PDFIO.PD.PDDoc — Type.

    PDDoc

A in memory representation of a PDF document. Once created this type has to be used to access a PDF document.

source

PDFIO.PD.pdDocOpen — Function.

    pdDocOpen(filepath::AbstractString) -> PDDoc

Opens a PDF document and provides the PDDoc document object for subsequent query into the PDF file. filepath is the path to the PDF file in the relative or absolute path format. Remember to release the document with pdDocClose, once the object is used.

source

PDFIO.PD.pdDocClose — Function.

    pdDocClose(doc::PDDoc, num::Int) -> PDDoc

Reclaim the resources associated with a PDDoc object. Once called the PDDoc object cannot be further used.

source

PDFIO.PD.pdDocGetCatalog — Function.

    pdDocGetCatalog(doc::PDDoc) -> CosObject

Catalog is considered the topmost level object in PDF document that is subsequently used to traverse and extract information on a PDF document. To be used for accessing PDF internal objects from document structure when no direct API is available.

source

PDFIO.PD.pdDocGetNamesDict — Function.

    pdDocGetNamesDict(doc::PDDoc) -> CosObject

Some information in PDF is stored as name and value pairs not essentially a dictionary. They are all aggregated and can be accessed from one names dictionary object in the document catalog. This method provides access to such dictionary in a PDF file. Not all PDF document may have a names dictionary. In such cases, a CosNull object may be returned.

source

PDFIO.PD.pdDocGetInfo — Function.

    pdDocGetInfo(doc::PDDoc) -> Dict

Given a PDF document provides the document information available in the DocumentInfo disctionary. The information typically includes _creation date, modification date, author, creator_ used etc. However, information content are not all mandatory and all information may not be available in a document. Please refer to the PDF specification for details.

source

PDFIO.PD.pdDocGetCosDoc — Function.

    pdDocGetCosDoc(doc::PDDoc) -> CosDoc

PDF document format is developed in two layers. A logical PDF document information is represented over a physical file structure called COS. CosDoc is an access object to the physical file structure of the PDF document. To be used for accessing PDF internal objects from document structure when no direct API is available.

One can access any aspect of PDF using the COS level APIs alone. However, they may require you to know the PDF specification in details and not the most intuititive.

source

PDFIO.PD.pdDocGetPage — Function.

    pdDocGetPage(doc::PDDoc, num::Int) -> PDPage

Given a document absolute page number provides the associated page.

source

PDFIO.PD.pdDocGetPageCount — Function.

    pdDocGetPageCount(doc::PDDoc) -> Int

Returns the number of pages associated with the document.

source

PDFIO.PD.pdDocGetPageRange — Function.

    pdDocGetPageRange(doc::PDDoc, nums::Range{Int}) -> Vector{PDPage}
    pdDocGetPageRange(doc::PDDoc, label::AbstractString) -> Vector{PDPage}

Given a range of page numbers or a label returns an array of pages associated with it.

source

PDFIO.PD.pdPageGetContents — Function.

    pdPageGetContents(page::PDPage) -> CosObject

Page rendering objects are normally stored in a CosStream object in a PDF file. This method provides access to the stream object.

source

PDFIO.PD.pdPageIsEmpty — Function.

    pdPageIsEmpty(page::PDPage) -> Bool

Returns true when the page has no associated content object.

source

PDFIO.PD.pdPageGetCosObject — Function.

    pdPageGetCosObject(page::PDPage) -> CosObject

PDF document format is developed in two layers. A logical PDF document information is represented over a physical file structure called COS. This method provides the internal COS object associated with the page object.

source

PDFIO.PD.pdPageGetContentObjects — Function.

    pdPageGetContentObjects(page::PDPage) -> CosObject

A page object can have multiple associated content objects. This method will return a CosArray if there are multiple of content objects associated. Otherwise an indirect object of CosStream type will be returned.

source

Cos

PDFIO.Cos.CosDoc — Type.

    CosDoc

PDF document format is developed in two layers. A logical PDF document information is represented over a physical file structure called COS. CosDoc is an access object to the physical file structure of the PDF document. To be used for accessing PDF internal objects from document structure when no direct API is available.

One can access any aspect of PDF using the COS level APIs alone. However, they may require you to know the PDF specification in details and not the most intuititive.

source

PDFIO.Cos.cosDocOpen — Function.

    cosDocOpen(filepath::String) -> CosDoc

Provides the access to the physical file and file structure of the PDF document. Returns a CosDoc which can be subsequently used for all query into the PDF files. Remember to release the document with cosDocClose, once the object is used.

source

PDFIO.Cos.cosDocClose — Function.

    cosDocClose(doc::CosDoc)

Reclaims all system resources consumed by the CosDoc. The CosDoc should not be used after this method is called. cosDocClose only needs to be explicitly called if you have opened the document by'cosDocOpen'. Documents opened with pdDocOpen do not need to use this method.

source

PDFIO.Cos.cosDocGetRoot — Function.

    cosDocGetRoot(doc::CosDoc) -> CosDoc

The structural starting point of a PDF document. Also known as document root dictionary. This provides details object locations and document access methodology. This should not be confused with the catalog object of the PDF document.

source

PDFIO.Cos.cosDocGetObject — Function.

    cosDocGetObject(doc::CosDoc, obj::CosObject) -> CosObject

PDF objects are distributed in the file and can be cross referenced from one location to another. This is called as indirect object referencing. However, to extract actual information one needs access to the complete object (direct object). This method provides access to the direct object after searching for the object in the document structure. If an indirect object reference is passed as an obj parameter the complete indirect object (reference as well as all content of the object) are returned. A direct object passed to the method is returned as is without any translation. This ensures the user does not have to go through type check of the objects before accessing the contents.

source

PDFIO.Cos.cosDocGetPageNumbers — Function.

cosDocGetPageNumbers(doc::CosDoc, catalog::CosObject, label::AbstractString) -> Range{Int}

PDF utilizes two pagination schemes. An internal global page number that is maintained serially as an integer and PageLabel that is shown by the viewers. Given a label this method returns a range of valid page numbers for the given label.

source